// asr.webapp.v2/src/app/modules/configuration/measurement/unit/unit.types.ts
export class Units {
    unitId: number;
    unitCode: string;
    unitName: string;
    unitType: number;
    measurementMetricId: number;
    measurementMetricName: string;
    status: number;
    languageID: number;
    clientIP: string;
    registerUserID: number;
    registerUserName: string;
    localChangeDate: string;
    localChangeTime: string;
    chaneDate: string;
    abbreviation: string;
}